+++ pixelfed/CHANGELOG.md
@@ -5,3 +5,4 @@
 * First bash at a pot jail for pixelfed
 * Fix typos
 * Fix env file formatting
+* Redis must be running before pixelfed is configured, optionally user www must be part of redis group